requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
amqp==5.3.1
    # via kombu
babel==2.17.0
    # via frontend-n8n-skripsi (pyproject.toml)
beautifulsoup4==4.13.4
    # via gdown
bidict==0.23.1
    # via python-socketio
billiard==4.2.1
    # via celery
blinker==1.9.0
    # via flask
boto3==1.40.48
    # via frontend-n8n-skripsi (pyproject.toml)
botocore==1.40.48
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.1
    # via google-auth
celery==5.5.3
    # via frontend-n8n-skripsi (pyproject.toml)
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
dnspython==2.8.0
    # via eventlet
eventlet==0.40.3
    # via frontend-n8n-skripsi (pyproject.toml)
filelock==3.18.0
    # via gdown
flask==3.1.1
    # via
    #   frontend-n8n-skripsi (pyproject.toml)
    #   flask-admin
    #   flask-socketio
    #   flask-sqlalchemy
flask-admin==2.0.0
    # via frontend-n8n-skripsi (pyproject.toml)
flask-socketio==5.5.1
    # via frontend-n8n-skripsi (pyproject.toml)
flask-sqlalchemy==3.1.1
    # via frontend-n8n-skripsi (pyproject.toml)
gdown==5.2.0
    # via frontend-n8n-skripsi (pyproject.toml)
gevent==25.9.1
    # via frontend-n8n-skripsi (pyproject.toml)
google-api-core==2.27.0
    # via google-api-python-client
google-api-python-client==2.185.0
    # via pydrive
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.71.0
    # via google-api-core
greenlet==3.2.4
    # via
    #   eventlet
    #   gevent
    #   sqlalchemy
gunicorn==23.0.0
    # via frontend-n8n-skripsi (pyproject.toml)
h11==0.16.0
    # via wsproto
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   flask-admin
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.5.4
    # via celery
markupsafe==3.0.2
    # via
    #   flask
    #   flask-admin
    #   jinja2
    #   werkzeug
    #   wtforms
numpy==2.3.1
    # via
    #   frontend-n8n-skripsi (pyproject.toml)
    #   pandas
oauth2client==4.1.3
    # via pydrive
packaging==25.0
    # via
    #   gunicorn
    #   kombu
pandas==2.3.0
    # via frontend-n8n-skripsi (pyproject.toml)
prompt-toolkit==3.0.51
    # via click-repl
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.11
    # via frontend-n8n-skripsi (pyproject.toml)
pyasn1==0.6.1
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via
    #   google-auth
    #   oauth2client
pydrive==1.3.1
    # via frontend-n8n-skripsi (pyproject.toml)
pyparsing==3.2.5
    # via httplib2
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   pandas
python-dotenv==1.1.1
    # via frontend-n8n-skripsi (pyproject.toml)
python-engineio==4.12.2
    # via python-socketio
python-socketio==5.13.0
    # via flask-socketio
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via pydrive
redis==6.2.0
    # via frontend-n8n-skripsi (pyproject.toml)
requests==2.32.4
    # via
    #   gdown
    #   google-api-core
rsa==4.9.1
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.14.0
    # via boto3
setuptools==80.9.0
    # via zope-event
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via
    #   oauth2client
    #   python-dateutil
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via
    #   frontend-n8n-skripsi (pyproject.toml)
    #   flask-sqlalchemy
tqdm==4.67.1
    # via gdown
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   sqlalchemy
tzdata==2025.2
    # via
    #   kombu
    #   pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
waitress==3.0.2
    # via frontend-n8n-skripsi (pyproject.toml)
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via
    #   flask
    #   flask-admin
wsproto==1.2.0
    # via simple-websocket
wtforms==3.2.1
    # via flask-admin
zope-event==6.0
    # via gevent
zope-interface==8.0.1
    # via gevent
